Renamed project to UnifiedASATVisualizer.
