+++ pom.xml
@@ -3,7 +3,7 @@
   <modelVersion>4.0.0</modelVersion>
 
   <groupId>BlueTurtle</groupId>
-  <artifactId>TSE</artifactId>
+  <artifactId>UnifiedASATVisualizer</artifactId>
   <version>1.1</version>
   <packaging>jar</packaging>
 
@@ -73,7 +73,7 @@
 				  <transformers>
 					<!-- add Main-Class to manifest file -->
    	                       <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-						   <mainClass>BlueTurtle.TSE.Main</mainClass>
+						   <mainClass>BlueTurtle.uav.Main</mainClass>
 				           </transformer>
 				  </transformers>
 				  <outputDirectory>${basedir}</outputDirectory>
