+++ src/main/java/BlueTurtle/finders/ProjectInfoFinder.java
@@ -7,9 +7,9 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Set;
 
-import BlueTurtle.TSE.CodeFile;
 import BlueTurtle.computers.LOCComputer;
 import BlueTurtle.gui.GUIController.ASAT;
+import BlueTurtle.uav.CodeFile;
 import BlueTurtle.writers.JSWriter;
 import lombok.Getter;
