+++ src/main/java/BlueTurtle/gui/GUIController.java
@@ -8,10 +8,10 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-import BlueTurtle.TSE.JavaController;
-import BlueTurtle.TSE.Main;
 import BlueTurtle.finders.ProjectInfoFinder;
 import BlueTurtle.gui.GUIController.ASAT;
+import BlueTurtle.uav.JavaController;
+import BlueTurtle.uav.Main;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
