+++ src/main/java/BlueTurtle/uav/CodeFile.java
@@ -1,4 +1,4 @@
-package BlueTurtle.TSE;
+package BlueTurtle.uav;
 
 import java.io.BufferedReader;
 import java.io.File;
