+++ src/main/java/BlueTurtle/uav/Main.java
@@ -1,4 +1,4 @@
-package BlueTurtle.TSE;
+package BlueTurtle.uav;
 
 import java.io.IOException;
