+++ src/main/java/BlueTurtle/writers/JSWriter.java
@@ -10,9 +10,9 @@ import java.util.List;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
-import BlueTurtle.TSE.CodeFile;
 import BlueTurtle.gui.GUIController;
 import BlueTurtle.summarizers.Summarizer;
+import BlueTurtle.uav.CodeFile;
 import lombok.Getter;
 import lombok.Setter;
